requirements.txt
aiogram
sqlalchemy
aiosqlite
apscheduler
python-dotenv
pyrogram
cryptg